Stable 2024.2

+++ hardware/S0MeterTCP.cpp
@@ -82,7 +82,7 @@ bool S0MeterTCP::WriteToHardware(const char *pdata, const unsigned char length)
 
 void S0MeterTCP::OnData(const unsigned char *pData, size_t length)
 {
-	ParseData((const unsigned char*)pData,length);
+	ParseData((const unsigned char*)pData,static_cast<int>(length));
 }
 
 void S0MeterTCP::OnError(const boost::system::error_code& error)
